Refactor two files
