+++ tests/configlet/util/base_test.py
@@ -2,16 +2,17 @@
 
 import json
 import os
+import configlet
 
-from helpers import set_log_prefix_msg, get_prefix_lvl, set_prefix_lvl, append_log_prefix_msg,\
+from tests.configlet.util import strip
+from tests.configlet.util import generic_patch
+from tests.common.configlet.helpers import set_log_prefix_msg, get_prefix_lvl, set_prefix_lvl, append_log_prefix_msg,\
                     log_info, log_debug
-from common import base_dir, data_dir, orig_db_dir, no_t0_db_dir, clet_db_dir, managed_files,\
-                   patch_add_t0_dir, patch_rm_t0_dir, files_dir, tor_data, init_data,\
+from tests.common.configlet.utils import base_dir, data_dir, orig_db_dir, no_t0_db_dir, clet_db_dir, managed_files,\
+                   patch_add_t0_dir, patch_rm_t0_dir, files_dir, tor_data, init_data, \
                    RELOAD_WAIT_TIME, PAUSE_INTF_DOWN, PAUSE_INTF_UP, PAUSE_CLET_APPLY, DB_COMP_WAIT_TIME,\
                    do_pause, db_comp, chk_bgp_session, chk_for_pfc_wd, report_error, take_DB_dumps, init_global_data
-import strip
-import configlet
-import generic_patch
+
 
 if os.path.exists("/etc/sonic/sonic-environment"):
     from mock_for_switch import config_reload, wait_until
