+++ tests/configlet/util/run_test_in_switch.py
@@ -8,7 +8,7 @@ import os
 from mock_for_switch import get_duthost
 
 from base_test import do_test_add_rack, backup_minigraph, restore_orig_minigraph
-from helpers import log_error, set_print
+from tests.common.configlet.helpers import log_error, set_print
 
 # To run test in switch:
 # Copy all files in this dir (tests/configlet/util) into switch
