+++ tests/configlet/util/strip.py
@@ -4,8 +4,8 @@ import json
 import sys
 import xml.etree.ElementTree as ET
 
-from helpers import log_info, log_debug
-from common import tor_data, config_db_data_orig, managed_files, report_error   # noqa F401
+from tests.common.configlet.helpers import log_info, log_debug
+from tests.common.configlet.utils import tor_data, config_db_data_orig, managed_files, report_error   # noqa F401
 from tempfile import mkstemp
 
 ns_val = "Microsoft.Search.Autopilot.Evolution"
